+++ NINA/Model/GuideStepsHistory.cs
@@ -32,12 +32,13 @@ namespace NINA.Model {
 
     public class GuideStepsHistory : BaseINPC {
 
-        public GuideStepsHistory(int historySize) {
+        public GuideStepsHistory(int historySize, GuiderScaleEnum scale, double maxY) {
             RMS = new RMS();
             PixelScale = 1;
             HistorySize = historySize;
-            GuideSteps = new AsyncObservableLimitedSizedStack<IGuideStep>(historySize);
-            MaxY = 4;
+            MaxY = maxY;
+            Scale = scale;
+            GuideSteps = new AsyncObservableLimitedSizedStack<HistoryStep>(historySize);
             MaxDurationY = 1;
         }
 
@@ -102,11 +103,11 @@ namespace NINA.Model {
             }
         }
 
-        private LinkedList<IGuideStep> overallGuideSteps = new LinkedList<IGuideStep>();
+        private LinkedList<HistoryStep> overallGuideSteps = new LinkedList<HistoryStep>();
 
-        private AsyncObservableLimitedSizedStack<IGuideStep> guideSteps;
+        private AsyncObservableLimitedSizedStack<HistoryStep> guideSteps;
 
-        public AsyncObservableLimitedSizedStack<IGuideStep> GuideSteps {
+        public AsyncObservableLimitedSizedStack<HistoryStep> GuideSteps {
             get {
                 return guideSteps;
             }
@@ -124,15 +125,15 @@ namespace NINA.Model {
             }
             set {
                 historySize = value;
-                RebuildGuideScaleList();
+                RebuildGuideHistoryList();
                 RaisePropertyChanged();
             }
         }
 
-        private void RebuildGuideScaleList() {
+        private void RebuildGuideHistoryList() {
             lock (lockObj) {
                 if (overallGuideSteps.Count > 0) {
-                    var collection = new LinkedList<IGuideStep>();
+                    var collection = new LinkedList<HistoryStep>();
                     var startIndex = overallGuideSteps.Count - historySize;
                     if (startIndex < 0) startIndex = 0;
                     RMS.Clear();
@@ -140,12 +141,14 @@ namespace NINA.Model {
                         var p = overallGuideSteps.ElementAt(i);
                         RMS.AddDataPoint(p.RADistanceRaw, p.DECDistanceRaw);
                         if (Scale == GuiderScaleEnum.ARCSECONDS) {
-                            p = ConvertStepToArcSec(p);
+                            p = p.AdjustPixelScale(PixelScale);
+                        } else {
+                            p = p.AdjustPixelScale(1);
                         }
                         collection.AddLast(p);
                     }
 
-                    GuideSteps = new AsyncObservableLimitedSizedStack<IGuideStep>(historySize, collection);
+                    GuideSteps = new AsyncObservableLimitedSizedStack<HistoryStep>(historySize, collection);
                     CalculateMaximumDurationY();
                 }
             }
@@ -176,17 +179,9 @@ namespace NINA.Model {
             }
             set {
                 scale = value;
-                RebuildGuideScaleList();
+                RebuildGuideHistoryList();
                 RaisePropertyChanged();
             }
-        }
-
-        private IGuideStep ConvertStepToArcSec(IGuideStep step) {
-            var newStep = step.Clone();
-            // only displayed values are changed, not the raw ones
-            newStep.RADistanceRawDisplay = newStep.RADistanceRaw * PixelScale;
-            newStep.DECDistanceRawDisplay = newStep.DECDistanceRaw * PixelScale;
-            return newStep;
         }
 
         public void Clear() {
@@ -195,12 +190,14 @@ namespace NINA.Model {
                 GuideSteps.Clear();
                 RMS.Clear();
                 MaxDurationY = 1;
+                HistoryStep.ResetIdProvider();
             }
         }
 
         public void AddGuideStep(IGuideStep step) {
             lock (lockObj) {
-                overallGuideSteps.AddLast(step);
+                var historyStep = HistoryStep.FromGuideStep(step, Scale == GuiderScaleEnum.PIXELS ? 1 : PixelScale);
+                overallGuideSteps.AddLast(historyStep);
 
                 if (GuideSteps.Count == HistorySize) {
                     var elementIdx = overallGuideSteps.Count - HistorySize;
@@ -212,11 +209,75 @@ namespace NINA.Model {
 
                 RMS.AddDataPoint(step.RADistanceRaw, step.DECDistanceRaw);
 
-                if (Scale == GuiderScaleEnum.ARCSECONDS) {
-                    step = ConvertStepToArcSec(step);
-                }
-                GuideSteps.Add(step);
+                GuideSteps.Add(historyStep);
                 CalculateMaximumDurationY();
+            }
+        }
+
+        internal void AddDitherIndicator() {
+            lock (lockObj) {
+                var dither = HistoryStep.GenerateDitherStep();
+                overallGuideSteps.AddLast(dither);
+
+                GuideSteps.Add(dither);
+            }
+        }
+
+        public class HistoryStep {
+            private static int idProvider = 1;
+            public int Id { get; private set; }
+            public double IdOffsetLeft { get => Id - 0.15; }
+            public double IdOffsetRight { get => Id + 0.15; }
+            public double RADistanceRaw { get; private set; }
+            public double RADistanceRawDisplay { get; private set; }
+            public double RADuration { get; private set; }
+            public double DECDistanceRaw { get; private set; }
+            public double DECDistanceRawDisplay { get; private set; }
+            public double DECDuration { get; private set; }
+            public double Dither { get; private set; } = double.NaN;
+
+            public static HistoryStep FromGuideStep(IGuideStep guideStep, double pixelScale) {
+                var id = idProvider++;
+                return new HistoryStep() {
+                    Id = id,
+                    RADistanceRaw = guideStep.RADistanceRaw,
+                    RADistanceRawDisplay = guideStep.RADistanceRaw * pixelScale,
+                    RADuration = guideStep.RADuration,
+                    DECDistanceRaw = guideStep.DECDistanceRaw,
+                    DECDistanceRawDisplay = guideStep.DECDistanceRaw * pixelScale,
+                    DECDuration = guideStep.DECDuration
+                };
+            }
+
+            public static HistoryStep GenerateDitherStep() {
+                var id = idProvider++;
+                return new HistoryStep() {
+                    Id = id,
+                    RADistanceRaw = 0,
+                    RADistanceRawDisplay = 0,
+                    RADuration = 0,
+                    DECDistanceRaw = 0,
+                    DECDistanceRawDisplay = 0,
+                    DECDuration = 0,
+                    Dither = 0.01
+                };
+            }
+
+            public HistoryStep AdjustPixelScale(double pixelScale) {
+                return new HistoryStep() {
+                    Id = this.Id,
+                    RADistanceRaw = this.RADistanceRaw,
+                    RADistanceRawDisplay = this.RADistanceRaw * pixelScale,
+                    RADuration = this.RADuration,
+                    DECDistanceRaw = this.DECDistanceRaw,
+                    DECDistanceRawDisplay = this.DECDistanceRaw * pixelScale,
+                    DECDuration = this.DECDuration,
+                    Dither = this.Dither
+                };
+            }
+
+            public static void ResetIdProvider() {
+                idProvider = 0;
             }
         }
     }