+++ NINA/Profile/Interfaces/IGuiderSettings.cs
@@ -30,6 +30,7 @@ namespace NINA.Profile {
         double DitherPixels { get; set; }
         bool DitherRAOnly { get; set; }
         GuiderScaleEnum PHD2GuiderScale { get; set; }
+        double MaxY { get; set; }
         int PHD2HistorySize { get; set; }
         int PHD2ServerPort { get; set; }
         string PHD2ServerUrl { get; set; }