+++ RELEASE_NOTES.md
@@ -95,6 +95,9 @@
 - Added `BAYERPAT`, `XBAYEROFF`, and `YBAYEROFF` keyword support in FITS and XISF files for color cameras. A menu under Options > Equipment > Camera allows the user to override the driver-specified Bayer pattern with a custom one
 - Meridian Flip now has an option to do an auto focus after a flip during sequencing
 - When the camera is not giving a ready signal for image download after exposuretime + 15 seconds the exposure will be canceled and skipped
+- Guider graph pauses during dithering to not show guide pulses during dithering and just during "real" guiding
+- Guider graph settings are now stored in profile and reloaded when application is started again
+- Guider graph now displays a triangle as indicator when a dither action was happening
 
 ## Bug fixes
 - Guiding was improperly stopped when performing AF at start of sequence, even if DisableGuiding option was false